feat(ux): Improve user experience

+++ Assets/Scripts/io/dto/NpcDto.cs
@@ -8,10 +8,10 @@ namespace io.dto
         [JsonProperty("firstName")] public string FirstName { get; set; }
         [JsonProperty("lastName")] public string LastName { get; set; }
         [JsonProperty("prefab")] public string Prefab { get; set; }
-        [JsonProperty("spawnX")] public int SpawnX { get; set; }
-        [JsonProperty("spawnY")] public int SpawnY { get; set; }
-        [JsonProperty("spawnZ")] public int SpawnZ { get; set; }
-        [JsonProperty("spawnRotation")] public int SpawnRotation { get; set; }
+        [JsonProperty("spawnX")] public float SpawnX { get; set; }
+        [JsonProperty("spawnY")] public float SpawnY { get; set; }
+        [JsonProperty("spawnZ")] public float SpawnZ { get; set; }
+        [JsonProperty("spawnRotation")] public float SpawnRotation { get; set; }
         [JsonProperty("characterPrompt")] public string CharacterPrompt { get; set; }
         [JsonProperty("tools")] public NpcToolDto[] Tools { get; set; }
         [JsonProperty("createdAt")] public string CreatedAt { get; set; }
